backend/requirements.txt
# 基础框架
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=0.21.0

# 数据处理
pandas>=1.3.0
openpyxl>=3.0.7

# 数据库
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.1
alembic>=1.12.1

# SQL解析
sqlparse>=0.4.4

# 缓存
redis>=4.5.1

# 异步任务
celery==5.3.6

# 图算法
networkx==2.8.8

# 工具库
python-jose==3.3.0
passlib==1.7.4
python-multipart>=0.0.5,<0.1.0
email-validator==2.1.0.post1
pyyaml==6.0.1

# LLM 和 AI 工具
litellm>=0.1.345
loguru>=0.7.0
tenacity>=8.2.2
langchain>=0.0.267
langchain-openai>=0.0.2
tiktoken>=0.5.1

# MCP集成 (可选依赖)
# websockets>=11.0.3  # 如需使用WebSocket连接方式，请取消注释

# 学术搜索
scholarly>=1.7.0
arxiv>=1.4.7

# 测试
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# 开发工具
black==23.12.1
isort==5.13.2
flake8==7.0.0